Feat/ Refactor : Quiz ON Going Page
Refactored some Code for beter UI-UX of the page
tasks done
- [x] Added Logic for Quiz page
- [x] Added Fallback text for major Components
- [x] Added readable File Structure
