+++ src/components/QuizONgoing/Quizgrid.css
@@ -14,6 +14,13 @@
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
+    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
+    sans-serif;
+}
+
+.quesmargin {
+  margin-inline: 50px;
 }
 
 .infobar {
