+++ src/pages/QuizOngoing/quizpage.js
@@ -1,8 +1,17 @@
 import React from 'react'
-import Quizgrid from '../../components/Quizgrid'
+import Quizgrid from '../../components/QuizONgoing/Quizgrid'
 const quizpage = () => {
+    const Testoptionarr = [
+        'option 1',
+        'option 2',
+        'option 3',
+        'option 4',
+    ];
+    const testquesText = `This is  Test Text - Question`;
+    const testCounterText = `7`;
+
     return (
-        <><Quizgrid /></>
+        <Quizgrid testquesText={testquesText} optionArray={Testoptionarr} counter={testCounterText} />
     )
 }
